Moved handling of scheduled `Routine` instances to `Publisher`
