+++ src/storage/grouping.rs
@@ -1,4 +1,4 @@
-use crate::action::{IOCommand, PublisherInstance, Routine};
+use crate::action::{IOCommand, PublisherInstance};
 use crate::builders::DeviceLogBuilder;
 use crate::errors::ErrorType;
 use crate::helpers::{check_results, Deferred};
@@ -40,10 +40,6 @@ pub struct Group {
 
     /// Container for `PublisherInstances`
     pub publishers: Vec<Deferred<PublisherInstance>>,
-    /// Container for active `Routine` instances.
-    ///
-    /// `check_scheduled()` is the interface to attempt execution and clear this container.
-    pub scheduled: Vec<Routine>,
 }
 
 impl Group {
@@ -80,27 +76,6 @@ impl Group {
         }
     }
 
-    /// Attempt to execute scheduled routines.
-    ///
-    /// Scheduled routines are generated by `PublisherInstance`, however, execution of this
-    /// function should not be correllated with polling interval since the publisher schedules
-    /// routines without regard to group interval. Therefore, this function should be executed as
-    /// often as possible to maintain accuracy to publisher intention.
-    /// 
-    /// Any routines executed by `Routine::attempt()` are cleared from the `scheduled` container.
-    pub fn check_scheduled(&mut self) {
-        let mut executed = Vec::default();
-        for (index, routine) in self.scheduled.iter().enumerate() {
-            if routine.attempt() {
-                executed.push(index);
-            }
-        }
-        // remove completed
-        for index in executed {
-            self.scheduled.remove(index);
-        }
-    }
-
     /// Initialized empty containers.
     ///
     /// Builder and setter functions should be used to populate containers.
@@ -112,7 +87,6 @@ impl Group {
         let outputs = <DeviceContainer<IdType>>::default();
         let logs = Vec::default();
         let publishers = Vec::default();
-        let scheduled = Vec::default();
 
         Self {
             name: String::from(name),
@@ -122,7 +96,6 @@ impl Group {
             inputs,
             outputs,
             publishers,
-            scheduled,
         }
     }
 
@@ -219,6 +192,13 @@ impl Group {
             false => {
                 create_dir_all(path).expect("Could not create root data directory");
             },
+        }
+    }
+
+    pub fn check_scheduled(&self) {
+        for _publisher in self.publishers.iter() {
+            let mut publisher = _publisher.try_lock().unwrap();
+            publisher.check_scheduled();
         }
     }
 }
